requirements.txt
# API Rate Guardian 依赖

pyyaml>=6.0
requests>=2.28.0
